refactor: move `ImputerStrategy` into same file as `Imputer`
