+++ src/safeds/data/tabular/transformation/_imputer.py
@@ -2,15 +2,62 @@ from __future__ import annotations
 
 import sys
 import warnings
+from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Any
 
+from safeds._utils import _structural_hash
 from safeds.data.tabular.containers import Table
 from safeds.data.tabular.transformation._table_transformer import TableTransformer
-from safeds.data.tabular.typing import ImputerStrategy
 from safeds.exceptions import NonNumericColumnError, TransformerNotFittedError, UnknownColumnNameError
 
 if TYPE_CHECKING:
     from sklearn.impute import SimpleImputer as sk_SimpleImputer
+
+
+class ImputerStrategy(ABC):
+    """
+    The abstract base class of the different imputation strategies supported by the `Imputer`.
+
+    This class is only needed for type annotations. Use the subclasses nested inside `Imputer.Strategy` instead.
+    """
+
+    @abstractmethod
+    def _augment_imputer(self, imputer: sk_SimpleImputer) -> None:
+        """
+        Set the imputer strategy of the given imputer.
+
+        Parameters
+        ----------
+        imputer:
+            The imputer to augment.
+        """
+
+    @abstractmethod
+    def __eq__(self, other: object) -> bool:
+        """
+        Compare two imputer strategies.
+
+        Parameters
+        ----------
+        other:
+            other object to compare to
+
+        Returns
+        -------
+        equals:
+            Whether the two imputer strategies are equal
+        """
+
+    def __hash__(self) -> int:
+        """
+        Return a deterministic hash value for this imputer strategy.
+
+        Returns
+        -------
+        hash:
+            The hash value.
+        """
+        return _structural_hash(self.__class__.__qualname__)
 
 
 class Imputer(TableTransformer):
