+++ src/safeds/data/tabular/typing/__init__.py
@@ -6,7 +6,6 @@ import apipkg
 
 if TYPE_CHECKING:
     from ._column_type import Anything, Boolean, ColumnType, Integer, Nothing, RealNumber, String
-    from ._imputer_strategy import ImputerStrategy
     from ._schema import Schema
 
 apipkg.initpkg(
@@ -15,7 +14,6 @@ apipkg.initpkg(
         "Anything": "._column_type:Anything",
         "Boolean": "._column_type:Boolean",
         "ColumnType": "._column_type:ColumnType",
-        "ImputerStrategy": "._imputer_strategy:ImputerStrategy",
         "Integer": "._column_type:Integer",
         "Nothing": "._column_type:Nothing",
         "RealNumber": "._column_type:RealNumber",
@@ -28,7 +26,6 @@ __all__ = [
     "Anything",
     "Boolean",
     "ColumnType",
-    "ImputerStrategy",
     "Integer",
     "Nothing",
     "RealNumber",
